+++ tests/quick_test.sh
@@ -2,3 +2,4 @@
 
 
 python -m pytest -m 'quick and not todo' -k 'not test_trainer_models and not test_trainer_device and not test_trainer_load_save'
+python -m pytest -m 'quick and not todo' test_trainer_load_save.py